refactor files related to file pkg
