+++ pkg/file/exist.go
@@ -1,0 +1,12 @@
+package file
+
+import (
+	"os"
+)
+
+func Exist(filename string) bool {
+	if _, err := os.Stat(filename); os.IsNotExist(err) {
+		return false
+	}
+	return true
+}
